pyproject.toml
[project]
name = "rag-builder"
version = "0.1.0"
requires-python = ">= 3.12"
dependencies = [
    "aws-cdk-lib>=2.220.0",
    "constructs>=10.4.2",
]
